Add script to bring workspace from another screens
This is tailored to the workflow of having one dominant screen, and
using the others purely for reference. It assumes you always want the
focused workspace on the primary screen, so all the commands are
designed to bring a workspace to the current screen (or swap with
another).

Without multiple screens, it will just behave like normal i3 and focus
workspaces unobtrusively.